+++ packages/automation-hooks/src/constants.ts
@@ -1,0 +1,3 @@
+import { PublicKey } from "@solana/web3.js";
+
+export const TASK_QUEUE = new PublicKey('H39gEszvsi6AT4rYBiJTuZHJSF5hMHy6CKGTd7wzhsg7')