+++ packages/automation-hooks/src/hooks/useAutomateHotspotClaims.ts
@@ -1,3 +1,4 @@
+import { getAssociatedTokenAddressSync } from '@solana/spl-token'
 import {
   cronJobKey,
   cronJobNameMappingKey,
@@ -8,7 +9,7 @@ import {
   entityCronAuthorityKey,
   init as initHplCrons,
 } from '@helium/hpl-crons-sdk'
-import { sendInstructionsWithPriorityFee } from '@helium/spl-utils'
+import { HNT_MINT, sendInstructionsWithPriorityFee } from '@helium/spl-utils'
 import {
   customSignerKey,
   init as initTuktuk,
@@ -26,10 +27,10 @@ import { useAsyncCallback } from 'react-async-hook'
 import { useCronJob } from './useCronJob'
 import { useTaskQueue } from './useTaskQueue'
 import { AnchorProvider } from '@coral-xyz/anchor'
+import { TASK_QUEUE } from '../constants'
+import { useAccount } from '@helium/account-fetch-cache-hooks'
 
 type Schedule = 'daily' | 'weekly' | 'monthly'
-
-const TASK_QUEUE = new PublicKey('H39gEszvsi6AT4rYBiJTuZHJSF5hMHy6CKGTd7wzhsg7')
 
 const getScheduleCronString = (schedule: Schedule) => {
   // Get current time and add 1 minute
@@ -132,6 +133,7 @@ const TASK_RETURN_ACCOUNT_SIZE = 0.01
 const MIN_RENT = 0.00089088
 const EST_TX_FEE = 0.000001
 const RECIPIENT_RENT = 0.00242208
+const ATA_RENT = 0.002039 * LAMPORTS_PER_SOL;
 
 export const useAutomateHotspotClaims = ({
   schedule,
@@ -180,6 +182,14 @@ export const useAutomateHotspotClaims = ({
   const { amount: pdaWalletSol } = useSolOwnedAmount(
     pdaWallet,
   )
+  const ata = useMemo(() => {
+    if (!wallet) return undefined
+    return getAssociatedTokenAddressSync(
+      HNT_MINT,
+      wallet,
+      true
+    )
+  }, [wallet])
 
   const crankFundingNeeded = useMemo(() => {
     const minCrankReward = taskQueue?.minCrankReward?.toNumber() || 10000
@@ -187,12 +197,14 @@ export const useAutomateHotspotClaims = ({
       duration * minCrankReward
     )
   }, [duration, totalHotspots, taskQueue])
+  const { account } = useAccount(ata)
   const pdaWalletFundingNeeded = useMemo(() => {
     const minCrankReward = taskQueue?.minCrankReward?.toNumber() || 10000
     return (
       (MIN_RENT * LAMPORTS_PER_SOL) +
+      (account ? 0 : ATA_RENT) +
       // Actual claim txs
-      duration * (minCrankReward + 5000) * (totalHotspots || 1) +
+      duration * 20000 * (totalHotspots || 1) +
       // Requeue transactions (5 queues per tx)
       duration * minCrankReward * Math.ceil((totalHotspots || 1) / 5)
     )