+++ packages/anchor-resolvers/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.13](https://github.com/helium/helium-program-library/compare/v0.10.12...v0.10.13) (2025-05-12)
+
+**Note:** Version bump only for package @helium/anchor-resolvers
+
+
+
+
+
+## [0.10.12](https://github.com/helium/helium-program-library/compare/v0.10.11...v0.10.12) (2025-05-05)
+
+**Note:** Version bump only for package @helium/anchor-resolvers
+
+
+
+
+
 ## [0.10.11](https://github.com/helium/helium-program-library/compare/v0.10.10...v0.10.11) (2025-04-30)
 
 **Note:** Version bump only for package @helium/anchor-resolvers